requirements.txt
kafka-python~=2.0.5
pyspark~=3.5.5
scikit-learn~=1.6.1
cassandra-driver~=3.29.2
gevent>=23.0.0
pyyaml~=6.0.2
numpy~=2.2.3
streamlit~=1.42.2
pandas~=2.2.3
